requirements.txt
Django==2.2.5
pandas==1.1.5
psycopg2-binary==2.8.6
djangorestframework==3.12.2
